requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
psycopg2-binary==2.9.10
openai==1.59.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
pymilvus>=2.3.0
protobuf>=4.21.0
grpcio>=1.49.0
